# apps/api/openapi.json
{
...
  "info": {
    "description": "Versioned transport contract for Airline Manager.",
    "title": "Airline Manager API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/airlines/{airlineId}/fuel/capacity-offers": {
      "get": {
        "operationId": "getFuelCapacityOffers",
        "parameters": [
          {
            "in": "path",
            "name": "airlineId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": false,
                    "properties": {
                      "capacityKg": {
                        "pattern": "^[1-9][0-9]*$",
                        "type": "string"
                      },
                      "currency": {
                        "enum": ["CHF", "EUR", "GBP", "JPY", "KWD", "USD"],
                        "type": "string"
                      },
                      "fuelRulesVersion": {
                        "type": "string"
                      },
                      "incrementalCapacityKg": {
                        "pattern": "^[1-9][0-9]*$",
                        "type": "string"
                      },
                      "priceMinor": {
                        "pattern": "^[1-9][0-9]*$",
                        "type": "string"
                      },
                      "tier": {
                        "minimum": 2,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "tier",
                      "capacityKg",
                      "incrementalCapacityKg",
                      "currency",
                      "priceMinor",
                      "fuelRulesVersion"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": ["fuel"]
      }
    },
...
    "/v1/airlines/{airlineId}/fuel/forecast": {
      "post": {
        "operationId": "forecastFuel",
        "parameters": [
          {
            "in": "path",
            "name": "airlineId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "projectedConsumptionKg": {
                    "pattern": "^[0-9]+$",
                    "type": "string"
                  }
                },
                "required": ["projectedConsumptionKg"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "advisoryOnly": {
                      "enum": [true],
                      "type": "boolean"
                    },
                    "airlineId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "minimumReserveKg": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "onHandKg": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "planningReservedKg": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "projectedAvailableKg": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "projectedConsumptionKg": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "projectedOnHandKg": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "projectedShortageKg": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "airlineId",
                    "onHandKg",
                    "planningReservedKg",
                    "minimumReserveKg",
                    "projectedConsumptionKg",
                    "projectedOnHandKg",
                    "projectedAvailableKg",
                    "projectedShortageKg",
                    "advisoryOnly"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "details": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": ["issue"],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message", "requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": ["fuel"]
      }
    },
...
  },
  "servers": [
    {
      "description": "Current origin",
      "url": "/"
    }
  ],
...
}